move TestEnv to spec helper
